requirements.txt
SQLAlchemy~=2.0.35
python-dotenv~=1.0.1